skills-store/pyproject.toml
[build-system]
requires = ["setuptools>=68.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "skills-store-backend"
version = "0.1.0"
description = "Skills Store MVP backend service"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "alembic>=1.14.0",
    "asyncpg>=0.30.0",
    "bcrypt>=4.2.0",
    "cryptography>=44.0.0",
    "fastapi>=0.115.0",
    "passlib[bcrypt]>=1.7.4",
    "psycopg[binary]>=3.2.0",
    "pydantic-settings>=2.8.0",
    "python-jose>=3.3.0",
    "python-multipart>=0.0.20",
    "pyyaml>=6.0.2",
    "sqlalchemy>=2.0.38",
    "uvicorn[standard]>=0.34.0",
]

[project.optional-dependencies]
dev = [
    "httpx>=0.28.0",
    "mypy>=1.15.0",
    "pytest>=8.3.0",
    "pytest-asyncio>=0.25.0",
    "ruff>=0.11.0",
]

[tool.setuptools.packages.find]
where = ["backend"]

[tool.ruff]
line-length = 100
target-version = "py311"

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py"]
